Update on Jasons checklist.
This checklist consists of the things My friend jaosn asked me to focus on to have a good understanding of the c_fundimentals, there are more things here to refet to, and more stiff to read but the knowledge is worth it. 

But this checklist has my understading on the things in c but the questions belong ot Jason.